# feat(spanner): add SelectAll method to decode from Spanner iterator.Rows to golang struct