--- a/spanner/read.go
+++ b/spanner/read.go
@@ -21,7 +21,6 @@ import (
 	"context"
 	"io"
 	"log"
-	"strconv"
 	"sync/atomic"
 	"time"
 
@@ -91,12 +90,11 @@ func streamWithReplaceSessionFunc(
 	}
 }
 
-// Iterator is an interface for iterating over Rows.
-type Iterator interface {
+// rowIterator is an interface for iterating over Rows.
+type rowIterator interface {
 	Next() (*Row, error)
 	Do(f func(r *Row) error) error
 	Stop()
-	RowsReturned() int64
 }
 
 // RowIterator is an iterator over Rows.
@@ -130,28 +128,7 @@ type RowIterator struct {
 	sawStats         bool
 }
 
-// RowsReturned returns, a lower bound on the number of rows returned by the query.
-// Currently, this requires the query to be executed with query stats enabled.
-//
-// If the query was a DML statement, the number of rows affected is returned.
-// If the query was a PDML statement, the number of rows affected is a lower bound.
-// If the query was executed without query stats enabled, or if it is otherwise
-// impossible to determine the number of rows in the resultset, -1 is returned.
-func (r *RowIterator) RowsReturned() int64 {
-	if r.sawStats && r.QueryStats != nil && r.QueryStats["rows_returned"] != nil {
-		switch rowsReturned := r.QueryStats["rows_returned"].(type) {
-		case float64:
-			return int64(rowsReturned)
-		case string:
-			v, err := strconv.ParseInt(rowsReturned, 10, 64)
-			if err != nil {
-				v = -1
-			}
-			return v
-		}
-	}
-	return -1
-}
+var _ rowIterator = (*RowIterator)(nil)
 
 // Next returns the next result. Its second return value is iterator.Done if
 // there are no more results. Once Next returns Done, all subsequent calls