… spanner/go.sum
… spanner/mocks.go
… spanner/row_test.go