@rahul2393 2023-12-29
@CAFxX Can you please help review this please, while this is not the final optimization but its better than using toStruct
Thanks