@CAFxX 2024-01-02
> this is not the final optimization

@rahul2393 to be able to know what to focus on in the review, can you give me a rough idea of what kind of optimizations you are imagining to do?